Создан сервис удаления объявления

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -1,6 +1,5 @@
 package ru.skypro.homework.repository;
 
-import org.springframework.cglib.transform.impl.AddInitTransformer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.dto.AdDto;
@@ -10,10 +9,13 @@ import ru.skypro.homework.model.User;
 import java.util.List;
 
 @Repository
-public interface AdRepository extends JpaRepository<Ad, Long> {
+public interface AdRepository extends JpaRepository<Ad, Integer> {
     List<Ad> findAdByUser(User user);
 
     Ad findByPk(int pk);
+
+//    void deleteById(int pk);
+
 //    Ad findAdByUser_Id(int pk);
 //
 //    AdDto deleteById(int pk);
